fix(openapi): add schema constraints — descriptions, string/int/array limits
- Add descriptions to all ~55 top-level schemas across 9 schema files
- Add maxLength + pattern/format/enum to all string properties (fixes owasp-string-limit, owasp-string-restricted)
- Add minimum/maximum + format to all integer properties (fixes owasp-integer-limit, owasp-integer-format)
- Add maxItems to all array properties (fixes owasp-array-limit)
- Add constraints to inline path/query parameters in all 9 path files
- Fix server URL to https (owasp-security-hosts-https-oas3)
- Add RFC8725 description to JWT security scheme (owasp-jwt-best-practices)
- Update Go handlers for new enum types and integer format changes

Reduces linting violations from 2,576 to 1,717.

+++ internal/api/handler_governance.go
@@ -195,9 +195,9 @@ func (h *APIHandler) DeleteLineageEdge(ctx context.Context, req DeleteLineageEdg
 
 // PurgeLineage implements the endpoint for purging lineage data older than a threshold.
 func (h *APIHandler) PurgeLineage(ctx context.Context, req PurgeLineageRequestObject) (PurgeLineageResponseObject, error) {
-	deleted, err := h.lineage.PurgeOlderThan(ctx, req.Body.OlderThanDays)
+	deleted, err := h.lineage.PurgeOlderThan(ctx, int(req.Body.OlderThanDays))
 	if err != nil {
-		code := errorCodeFromError(err)
+		code := int32(errorCodeFromError(err))
 		return PurgeLineage403JSONResponse{Code: code, Message: err.Error()}, nil
 	}
 	return PurgeLineage200JSONResponse{DeletedCount: &deleted}, nil
